requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.20

# LangChain Core
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5

# OpenAI
openai==1.82.0

# Vector Store & Embeddings
faiss-cpu==1.7.4
numpy==1.24.3

# Database
pymongo==4.6.1

# Document Processing
unstructured==0.17.2
unstructured-client==0.35.0
pypdf==5.5.0
python-docx==1.1.2

# Text Processing
nltk==3.9.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic_core==2.14.1

# Development & Testing
pytest==7.4.3
httpx==0.25.2

# Core Dependencies
requests==2.32.3
aiofiles==24.1.0
typing_extensions==4.13.2
